refactor: move rake tasks test helpers from core to dev to reuse them
