+++ decidim-core/spec/tasks/decidim_tasks_check_users_newsletter_opt_in_spec.rb
@@ -1,12 +1,8 @@
 # frozen_string_literal: true
 
 require "spec_helper"
-require "support/tasks"
-require "tasks/decidim_tasks_sniffs_stdout_context"
 
 describe "rake decidim:check_users_newsletter_opt_in", type: :task do
-  include_context "with stdout sniffing"
-
   let!(:user_wo) { create(:user, :confirmed, newsletter_notifications_at: nil) }
   let!(:user_w_optin) { create(:user, :confirmed, newsletter_notifications_at: Time.zone.parse("2018-05-26 00:00 +02:00")) }
   let!(:users) { create_list(:user, 4, :confirmed, newsletter_notifications_at: Time.zone.parse("2018-05-24 00:00 +02:00")) }
