+++ decidim-core/spec/tasks/decidim_tasks_delete_data_portability_files_spec.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 require "spec_helper"
-require "support/tasks"
-require "tasks/decidim_tasks_sniffs_stdout_context"
 
 describe "rake decidim:delete_data_portability_files", type: :task do
   let!(:original_expiry_time) { Decidim.data_portability_expiry_time }
